Remove tests again and re-enable bits that should now work
